Further restructuring; routes defined with prefixes
