+++ backend/app/models/email_model.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel, Field, ConfigDict
+
+class Email(BaseModel):
+    id: int
+    from_: str = Field(alias='from')
+    subject: str
+    body: str
+    model_config = ConfigDict(populate_by_name=True)
