+++ backend/app/models/summary_model.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel, Field, ConfigDict
+
+class EmailSummary(BaseModel):
+    email_id: int
+    summary: str
