+++ backend/app/services/auth_service.py
@@ -1,0 +1,53 @@
+
+import os
+import pickle
+from google_auth_oauthlib.flow import InstalledAppFlow
+from google.auth.transport.requests import Request
+from googleapiclient.discovery import build
+from dotenv import load_dotenv
+
+load_dotenv()
+
+SCOPES = ['https://mail.google.com/']
+
+
+def get_credentials():
+    creds = None
+    token_pickle = 'token.pickle'
+
+    # Load existing credentials from token.pickle, if they exist
+    if os.path.exists(token_pickle):
+        with open(token_pickle, 'rb') as token:
+            creds = pickle.load(token)
+
+    # If there are no valid credentials available, let the user log in.
+    if not creds or not creds.valid:
+        if creds and creds.expired and creds.refresh_token:
+            creds.refresh(Request())
+        else:
+            # Load client ID and client secret from environment variables
+            client_id = os.getenv('GOOGLE_CLIENT_ID')
+            client_secret = os.getenv('GOOGLE_CLIENT_SECRET')
+
+            if not client_id or not client_secret:
+                raise Exception("Google API credentials not found in environment variables.")
+
+            # Construct the client configuration
+            client_config = {
+                "installed": {
+                    "client_id": client_id,
+                    "client_secret": client_secret,
+                    "redirect_uris": ["urn:ietf:wg:oauth:2.0:oob", "http://localhost"],
+                    "auth_uri": "https://accounts.google.com/o/oauth2/auth",
+                    "token_uri": "https://oauth2.googleapis.com/token"
+                }
+            }
+
+            flow = InstalledAppFlow.from_client_config(client_config, SCOPES)
+            creds = flow.run_local_server(port=0)
+
+        # Save the credentials for the next run
+        with open(token_pickle, 'wb') as token:
+            pickle.dump(creds, token)
+
+    return creds
